Fix compilation on ia64

+++ src/skins/vxworks/taskLib.c
@@ -393,5 +393,5 @@ TASK_ID taskNameToId(const char *name)
 
 void taskExit(int code)
 {
-	pthread_exit((void *) code);
+	pthread_exit((void *)(long) code);
 }
